Rename from "new" job manager
