+++ server/src/api/modules/competitions/competition.events.ts
@@ -1,5 +1,5 @@
-import newJobManager from '../../../jobs-new/job.manager';
-import { JobPriority } from '../../../jobs-new/job.utils';
+import jobManager from '../../../jobs/job.manager';
+import { JobPriority } from '../../../jobs/job.utils';
 import prisma, { Competition, Participation } from '../../../prisma';
 import { CompetitionWithParticipations, PlayerType } from '../../../utils';
 import * as discordService from '../../services/external/discord.service';
@@ -20,7 +20,7 @@ async function onParticipantsJoined(participations: Pick<Participation, 'playerI
   // Request updates for any new players
   players.forEach(({ username, type, registeredAt }) => {
     if (type !== PlayerType.UNKNOWN || Date.now() - registeredAt.getTime() > 60_000) return;
-    newJobManager.add('UpdatePlayerJob', { username });
+    jobManager.add('UpdatePlayerJob', { username });
   });
 }
 
@@ -42,7 +42,7 @@ async function onCompetitionStarted(competition: Competition) {
     discordService.dispatchCompetitionStarted(competition);
   });
 
-  newJobManager.add('UpdateCompetitionScoreJob', { competitionId: competition.id }, { skipDedupe: true });
+  jobManager.add('UpdateCompetitionScoreJob', { competitionId: competition.id }, { skipDedupe: true });
 }
 
 async function onCompetitionEnded(competition: Competition) {
@@ -54,7 +54,7 @@ async function onCompetitionEnded(competition: Competition) {
     discordService.dispatchCompetitionEnded(competitionDetails);
   });
 
-  newJobManager.add('UpdateCompetitionScoreJob', { competitionId: competition.id }, { skipDedupe: true });
+  jobManager.add('UpdateCompetitionScoreJob', { competitionId: competition.id }, { skipDedupe: true });
 }
 
 async function onCompetitionStarting(competition: Competition, period: EventPeriodDelay) {
@@ -79,7 +79,7 @@ async function onCompetitionEnding(competition: Competition, period: EventPeriod
 
     competitionDetails.participations
       .filter(p => p.progress.gained > 0) // Only update players that have gained xp
-      .forEach(p => newJobManager.add('UpdatePlayerJob', { username: p.player.username }));
+      .forEach(p => jobManager.add('UpdatePlayerJob', { username: p.player.username }));
 
     return;
   }
@@ -100,7 +100,7 @@ async function onCompetitionEnding(competition: Competition, period: EventPeriod
         return !p.player.updatedAt || Date.now() - p.player.updatedAt.getTime() > 1000 * 60 * 60 * 24;
       })
       .forEach(p => {
-        newJobManager.add('UpdatePlayerJob', { username: p.player.username }, { priority: JobPriority.LOW });
+        jobManager.add('UpdatePlayerJob', { username: p.player.username }, { priority: JobPriority.LOW });
       });
   }
 }
