+++ server/src/api/modules/groups/group.events.ts
@@ -1,4 +1,4 @@
-import newJobManager from '../../../jobs-new/job.manager';
+import jobManager from '../../../jobs/job.manager';
 import prisma from '../../../prisma';
 import { MemberJoinedEvent, MemberLeftEvent, MemberRoleChangeEvent, PlayerType } from '../../../utils';
 import * as discordService from '../../services/external/discord.service';
@@ -7,11 +7,11 @@ import { addToGroupCompetitions } from '../competitions/services/AddToGroupCompe
 import { removeFromGroupCompetitions } from '../competitions/services/RemoveFromGroupCompetitionsService';
 
 async function onGroupUpdated(groupId: number) {
-  newJobManager.add('UpdateGroupScoreJob', { groupId }, { skipDedupe: true });
+  jobManager.add('UpdateGroupScoreJob', { groupId }, { skipDedupe: true });
 }
 
 async function onGroupCreated(groupId: number) {
-  newJobManager.add('UpdateGroupScoreJob', { groupId }, { skipDedupe: true });
+  jobManager.add('UpdateGroupScoreJob', { groupId }, { skipDedupe: true });
 }
 
 async function onMembersRolesChanged(events: MemberRoleChangeEvent[]) {
@@ -45,10 +45,10 @@ async function onMembersJoined(events: MemberJoinedEvent[]) {
   // Request updates for any new players
   players.forEach(({ username, type, registeredAt }) => {
     if (type !== PlayerType.UNKNOWN || Date.now() - registeredAt.getTime() > 60_000) return;
-    newJobManager.add('UpdatePlayerJob', { username });
+    jobManager.add('UpdatePlayerJob', { username });
   });
 
-  newJobManager.add('UpdateGroupScoreJob', { groupId }, { skipDedupe: true });
+  jobManager.add('UpdateGroupScoreJob', { groupId }, { skipDedupe: true });
 }
 
 async function onMembersLeft(events: MemberLeftEvent[]) {
@@ -65,7 +65,7 @@ async function onMembersLeft(events: MemberLeftEvent[]) {
     await discordService.dispatchMembersLeft(groupId, playerIds);
   });
 
-  newJobManager.add('UpdateGroupScoreJob', { groupId }, { skipDedupe: true });
+  jobManager.add('UpdateGroupScoreJob', { groupId }, { skipDedupe: true });
 }
 
 export { onGroupCreated, onGroupUpdated, onMembersJoined, onMembersLeft, onMembersRolesChanged };
