+++ server/src/api/modules/groups/services/UpdateAllMembersService.ts
@@ -1,4 +1,4 @@
-import newJobManager from '../../../../jobs-new/job.manager';
+import jobManager from '../../../../jobs/job.manager';
 import prisma, { Player } from '../../../../prisma';
 import { Period, PeriodProps } from '../../../../utils';
 import { BadRequestError, NotFoundError } from '../../../errors';
@@ -20,7 +20,7 @@ async function updateAllMembers(groupId: number): Promise<number> {
 
   // Schedule an update job for every member
   for (const player of outdatedPlayers) {
-    newJobManager.add('UpdatePlayerJob', { username: player.username });
+    jobManager.add('UpdatePlayerJob', { username: player.username });
   }
 
   return outdatedPlayers.length;
