+++ server/src/api/modules/name-changes/name-change.events.ts
@@ -1,8 +1,8 @@
 import { NameChange } from '../../../prisma';
-import newJobManager from '../../../jobs-new/job.manager';
+import jobManager from '../../../jobs/job.manager';
 
 async function onNameChangeSubmitted(nameChange: NameChange) {
-  newJobManager.add('ReviewNameChangeJob', { nameChangeId: nameChange.id });
+  jobManager.add('ReviewNameChangeJob', { nameChangeId: nameChange.id });
 }
 
 export { onNameChangeSubmitted };
