+++ server/src/api/modules/players/player.events.ts
@@ -1,4 +1,4 @@
-import newJobManager from '../../../jobs-new/job.manager';
+import jobManager from '../../../jobs/job.manager';
 import { Player, Snapshot } from '../../../prisma';
 import { FlaggedPlayerReviewContext, PlayerType } from '../../../utils';
 import * as discordService from '../../services/external/discord.service';
@@ -43,8 +43,8 @@ async function onPlayerNameChanged(player: Player, previousDisplayName: string) 
   });
 
   // Setup jobs to assert the player's account type and auto-update them
-  newJobManager.add('UpdatePlayerJob', { username: player.username });
-  newJobManager.add('CheckPlayerTypeJob', { username: player.username });
+  jobManager.add('UpdatePlayerJob', { username: player.username });
+  jobManager.add('CheckPlayerTypeJob', { username: player.username });
 }
 
 async function onPlayerUpdated(
