+++ server/src/jobs/instances/AutoUpdatePatronPlayersJob.ts
