+++ server/src/jobs/instances/CheckPlayerTypeJob.ts
