+++ server/src/jobs/instances/ScheduleCompetitionEventsJob.ts
