+++ server/src/jobs/instances/ScheduleCompetitionScoreUpdatesJob.ts
