+++ server/src/jobs/instances/UpdateCompetitionScoreJob.ts
