+++ server/src/jobs/instances/UpdateGroupScoreJob.ts
